Fixed problem with fomatting of float point value of lat and lon. (value was missing point and made incorrect request, also it has an exponent sign)

+++ Zontik/Item.cs
@@ -9,7 +9,7 @@ namespace Zontik
     class Item
     {
         public string City { get; set; }
-        public double Lon { get; set; }
-        public double Lat { get; set; }
+        public decimal Lon { get; set; }
+        public decimal Lat { get; set; }
     }
 }
